Rename the module to wagtail_polymath
