+++ pyproject.toml
@@ -3,14 +3,14 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "wagtailmath"
+name = "wagtail-polymath"
 authors = [{name = "James Ramm", email = "jamessramm@gmail.com"}]
 maintainers = [{name = "Wagtail Nest team", email = "hello@wagtail.org"}]
 description = "Wagtail StreamField block for rendering mathematical equations"
 readme = "README.md"
 license = {file = "LICENSE"}
 classifiers = [
-    "Development Status :: 3 - Alpha",
+    "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: BSD License",
     "Operating System :: OS Independent",
@@ -53,7 +53,7 @@ Changelog = "https://github.com/wagtail-nest/wagtail-polymath/blob/main/CHANGELO
 Documentation = "https://github.com/wagtail-nest/wagtail-polymath/blob/main/docs/"
 
 [tool.flit.module]
-name = "wagtailmath"
+name = "wagtail_polymath"
 
 [tool.flit.sdist]
 exclude = [
