+++ src/wagtail_polymath/static/wagtail_polymath/js/wagtail_polymath.js
